Formatting document
Added text to elaborate on the steps to tidy data, along with descriptions of the original data variables and what they represent. This helps in structuring a rough outline in how ill approach the question. I also added a  description of the model that will be used for predictions later on.